feat: maintenance changes (#65)
* feat: maintenance changes, add log stacktrace, engine exceptions changed, engine timeout raise

* ref: unexpected/unknown exception message update

+++ src/main/java/pl/poznan/put/rnapdbee/backend/infrastructure/scheduler/ExpirationScheduler.java
@@ -58,7 +58,11 @@ public class ExpirationScheduler {
             if ((int) Duration.between(analysis.getCreatedAt(), Instant.now()).toDays() >= documentStorageDays) {
                 UUID expiredAnalysisId = analysis.getId();
                 expiredAnalysisDataIds.add(expiredAnalysisId);
-                analyzedFileService.deleteAnalyzedFile(expiredAnalysisId);
+
+                Boolean expiredAnalysisUsePdb = analysis.getUsePdb();
+                if (expiredAnalysisUsePdb == null || !expiredAnalysisUsePdb) {
+                    analyzedFileService.deleteAnalyzedFile(expiredAnalysisId);
+                }
 
                 switch (analysis.getScenario()) {
                     case SCENARIO_2D:
